Support autocomplete

+++ include/CLI/TypeTools.hpp
@@ -1544,7 +1544,7 @@ inline std::string sum_string_vector(const std::vector<std::string> &values) {
     } else {
         if(val <= static_cast<double>(std::numeric_limits<std::int64_t>::min()) ||
            val >= static_cast<double>(std::numeric_limits<std::int64_t>::max()) ||
-           val == static_cast<std::int64_t>(val)) {
+           val == static_cast<double>(static_cast<std::int64_t>(val))) {
             output = detail::value_string(static_cast<int64_t>(val));
         } else {
             output = detail::value_string(val);
